+++ VacanzeApi/Persistence/Repository/Grupo8/CruiserRepository.cs
@@ -112,7 +112,7 @@ namespace vacanze_back.VacanzeApi.Persistence.Repository.Grupo8
                     var id = Convert.ToInt32(table.Rows[i][0]);
                     var departureDate = Convert.ToString(table.Rows[i][2]);
                     var arrivalDate = Convert.ToString(table.Rows[i][3]);
-                    var price = Convert.ToDouble(table.Rows[i][4]);
+                    var price = Convert.ToDecimal(table.Rows[i][4]);
                     var locDeparture = Convert.ToInt32(table.Rows[i][5]);
                     var locArrival = Convert.ToInt32(table.Rows[i][6]);
                     Layover layover = new Layover(id,cruiserId,departureDate, arrivalDate, price, locDeparture, locArrival);
@@ -160,7 +160,7 @@ namespace vacanze_back.VacanzeApi.Persistence.Repository.Grupo8
                     var shipid = Convert.ToInt32(table.Rows[i][1]);
                     var departureDate = Convert.ToString(table.Rows[i][2]);
                     var arrivalDate = Convert.ToString(table.Rows[i][3]);
-                    var price = Convert.ToDouble(table.Rows[i][4]);
+                    var price = Convert.ToDecimal(table.Rows[i][4]);
                     var locDeparture = Convert.ToInt32(table.Rows[i][5]);
                     var locArrival = Convert.ToInt32(table.Rows[i][6]);
                     Layover layover = new Layover(id, shipid, departureDate, arrivalDate, price, locDeparture, locArrival);
